requirements.txt
numpy
Pillow
piexif

torch
torchvision

fastapi
uvicorn
typing-extensions
python-multipart